2188: extract DeleteEnrichmentButton from EnrichmentForm
